Add working loopnest example
